Rename metrics client to google cloud client
