+++ pkg/googlecloud/interface.go
@@ -1,0 +1,5 @@
+package googlecloud
+
+type GoogleCloudClient interface {
+	GetMetrics() (int32, error)
+}
