--- pkg/metrics/interface.go
@@ -1,5 +1,0 @@
-package metrics
-
-type MetricsClient interface {
-	GetMetrics() (int32, error)
-}
